refactor : #129 헤더 및 헤더 styled-components 분리
